Move static routes to pkg/static
