+++ pkg/static/svc.go
@@ -1,6 +1,8 @@
 package static
 
 import (
+	"net/http"
+
 	"github.com/go-chi/chi"
 )
 
@@ -9,4 +11,11 @@ var Svc Static
 type Static struct{}
 
 func (Static) Register(mux chi.Router) {
+	static := func(path string, contentType string, content []byte) {
+		mux.Get(path, func(w http.ResponseWriter, r *http.Request) {
+			w.Header().Set("Content-Type", contentType)
+			w.Write(content)
+		})
+	}
+	static(AppJS, "text/javascript", []byte(appJS))
 }
